Fix default family id field type to bigint

+++ db/migrate/20180310215241_add_default_family_to_user.rb
@@ -1,6 +1,6 @@
 class AddDefaultFamilyToUser < ActiveRecord::Migration[5.1]
   def change
-    add_column :users, :default_family_id , :integer
+    add_column :users, :default_family_id , :bigint
     add_foreign_key :users, :families, column: :default_family_id
   end
 end
